requirements.txt
asgiref==3.7.2
boto3==1.34.101
botocore==1.34.101
chardet==5.2.0
cloudinary==1.40.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.24
django-allauth==0.50.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-newsletter==0.9.1
django-newsletter-subscription==0.5.0
django-storages==1.14.3
easy-thumbnails==2.8.5
gunicorn==21.2.0
install==1.3.5
jmespath==1.0.1
ldif3==3.1.1
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python-card-me==0.9.3
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
s3transfer==0.10.1
sorl-thumbnail==12.10.0
sqlparse==0.4.4
stripe==9.4.0
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.26.18
